pyproject.toml
[tool.poetry]
name = "riic-engine"
version = "0.1.0"
description = "Standalone desktop app for simulating performance of RIIC setups"
authors = ["astral4 <88992929+astral4@users.noreply.github.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/ark-db/riic-engine"
packages = [{include = "riic_engine"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.1"
pillow = "^9.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
